Add test program for selection sort
The test program will print the time in ms that a recursive and an
iterative implementation of selection sort take to sort an array of
200000 random and reverse ordered integers. It also tests the ordered data
checking it against the result of `qsort()` from `stdlib.h`.

The implementations use C preprocessor macros to define the data type
and the comparison rules used for sorting.