Refactor to remove code duplication
The repeated lines of code used to define and execute a test were
substituted by a `struct timed_test` which holds:
* the test name;
* a function pointer to hold the choosen implementation, currently
recursive and iterative selection sort;
* fields to store the arguments to the function pointer;
* fields to store the ordered array and the expected result calculated
bythe standard function `qsort()`;
* fields to hold thhe size of the data to be ordered in bytes and the
size of each item;
* a status flag to hold the result of the comparison between the
expected and the ordered array;
* a field to recor a `struct timespec` to record the elapsed interval
taken to order the input data.

The `sizeof` operator is now used specifying the required size relative
to the element that will be stored. Previously, the type whose size was
requested was hardcoded, which might lead to future bugs.

The multiple array size calculations were replaced by a macro. As a
consequence of this, the hardcoded size limit of the array of pointer to
integer pointers `arrays` was removed.

The tests were defined in an array of `struct timed_test`. Each test is
processed by a function that replaces the duplicated code responsible
for initialising the arrays before the test, timing the execution of the
required sorting function implementation and comparing the obtained
result against the reference implementation.