Add bubble sort without nested loops
Add to the timing tests an implementation of bubble sort without the
nested `for` loops (`bubble_sort_optim`). One flag controls when the
sorting stops, and it is based on the property that one element is
always "floated" in its correct position on each iteraction until all
the elements are sorted in the array.