Add optimised bubble sort
The execution time of an optimised bubble sort implementation when
sorting random and reverse arrays will be tested. This implementation
tracks the items that were already placed in their correct positions
in tha dataset. In a bubble sort, at least one element will be on
its correct position on each iteration.

The descriptions of the bubble sort functions have been updated, as well
as their titles in the program output.