Add quicksort optimised with median of 3 pivot
Add to the tests an implementation of quicksort with the median of 3
pivot. The partition function still chooses the last element as the
pivot, but this element is now the median of 3 of the first, middle and
last elements of the array.

This implementation is also tested on reverse and random integer arrays.