Add median 3 pivot and shortest tail quicksort
Although GCC already optimises the tail recursion on quicksort, this new
implementation optimises the tail recursion so that the shortest tail is
followed in each iteration/recursion level.

There are no changes in the tests and the pivot is still selected from
the median of the first, middle and last elements in the integer array.