Remove macros from directly included files
A file for boilerplate macros used on the int arrays sorting
process has been created. Some macros were changed to better suit the
coding style and to ease the inference of their purposes.

This file is now included only in the implementation of the modules
(bubble, insertion, quick, selection and shell sort) in order to avoid
their inclusion on parts of the code that include the function
prototypes.

The function prototypes have been altered to accept void pointers as
their input. The correct type is casted inside each function using the
`ITEM` macro, effectively casting an int array from a void pointer.