Allocate output data for each test result
Remove the output int pointer field used for the definition of each test
routine. The allocation of the output data is made internally to the
`run_timed_tests`function for each test processed.

Remove the unnecessary arrays and update the initialisation of the test
array due to the field removal.

Update function descriptions and documentations.