Add timed tests with sorted arrays
Time the execution of the tested algorithms using a sorted array of ints
as input data. The following tests are now executed and timed with
inputs of random, reverse and sorted arrays:
* selection sort: iterative and recursive;
* bubble sort iterative: recursive, iterative without nested for loops
and iterative optimised;
* shell sort: iterative with h=3*h + 1 sequence;
* insertion sort: iterative and iterative with pivot;
* quicksort: recursive, recursive with median of 3, recursive with
median of 3 and shortest tail optimisation.

The standard recursive quicksort uses an enourmous amount of stack in
the ordered input test, sufficient for causing a segmentation fault on
systems with 8192 kiB of stack. Raising the stack size with `ulimit -s`
is advised (tests ran successfully with 16384 kiB of stack limit).

The order of selection sort tests has been modified to match the order
of the other algorithms' tests.