Count swaps, comparisons and recursive calls
Rewrite tests, test execution functions, test configurations and swap
and comparison macros in order to measure how many swap operations,
comparisons and recursive calls are executed for each evaluated
algorithm.

These new counters are defined as structs. These structs are created for
each algorithm and its auxiliar function located in a translation unit.
The configuration of the correct variable for each function is done by a
hidden pointer initialised by a macro.

The new macros depend on GCC extensions to update the number of
comparison operations.

For algorithms implemented in multiple functions (quicksort and shell
sort, for example) the number of recursive calls will track only the
calls from the main function. The swap and comparison operations
performed by the different functions will be added.

The reading and initialisation of the counters depend on two functions
that must exist in their translation unit: ine for clearing the global
structs and one for reading them. The correct struct is selected using
the address of each function.

Modify the Makefile in order for the compiler not issue warnings about
the use of GCC extensions and to speed up the build process.

Update comments and debug functions to redirect the correct messages to
stdout and keep the noisy (although improved) log output on stderr.

Increase the portability by changing the `printf()` format string used
for printing `size_t` variables to its correct value of `%zu`.

Correct a bug in quicksort's partition function that led to dangerous
macro expansion.

Due to quicksort's insane stack consumption on worse case scenarios and
the inclusion of the extra logging macros, the minimum stack size has
been increased to 32768 KiB (32 MiB). Executing `ulimit -Ss 32768`
before running the executable is stringly advised to avoid a segfault.