Create modules for each algorithm
The algorithms that were previously aggregated have been separated so
that each implementation has their own module/translation unit. The old
files used to aggregated the variations for each algorithm were deleted
and only their original headers remained in order to reduce the
necessary headers to be included in `main.c`.

The test interface has changed, and now each algorithm defines its own
function that will configure the test name, implementation details and
the functions used to sort the input and to clear and read the
performance counters.

Changes in the struct that registers the test, its inputs and execution
results demanded changes in the functions used to execute the tests, in
the reporting/printing routines and in the test definitions.