Add script for configuring and running the tests
The added shell script will assist in the execution of the compiled
binary and the report generation based on the output of the tests.

It sets up the recommended stack limit and records the output of the
tests in the standard output and standard error streams into files. The
total execution time is printed in the terminal after the exection.

The files keep the output of the test program and the timing difference
between the messages. Also, the files are updated on each newline
printed, and the result of the execution remains available in the
terminal.

The availability of the dependencies is not checked by the script, and
they are the following:
* the compiled binary (`sort_tests.elf`)
* GNU date
* GNU dc
* script (util-linux)
* ulimit (Bash)