Removed indirectly accessed functions from headers
The functions used to clear and read the performance counters were not
being accessed directly. As their exposure in the included headers was
unnecessary, they were removed from the public interface and moved to an
internal one, into the implementation files.

These functions are now private to their translation units (`static`
functions) and are made available through the functions responsible for
setting up the `struct timed_test`, the `get_defs` functions.